// packages/core/types.d.ts
export interface VerifyResult {
    /** Verification status */
    status: 'ok' | 'failed' | 'error';
    /** Verification timestamp */
    timestamp: number;
}

export interface InviteOptions {
    /** Expiration time */
    expire: number;
    /** Share text */
    text?: string;
    /** Suffix */
    suffix?: string;
    /** Payload */
    payload?: string;
    /** Start param */
    start_param?: string;
}

/**
 * Invite Result
 */
export interface InviteResult {
    /** Invite URL */
    url: string;
    /** Invite ID */
    key?: string;
}

/**
 * Invite Payload Result
 */
export interface InvitePayloadResult {
    /** Invite payload */
    payload: string;
}
// ...
export interface PortexResponse<T> {
    ok: boolean;
    body: PortexResponseBody<T> | null;
    status: number;
    statusText: string;
    headers: Headers;
}

export interface PortexResponseBody<T> {
    ok: boolean;
    result: T | null;
    is_dev?: boolean;
    desc?: string;
}

export interface PortexRequestOptions {
    method?: 'GET' | 'POST' | 'PUT' | 'DELETE';
    data?: Record<string, any>;
    headers?: Record<string, string>;
}

/**
 * Request Interface
 */
export interface IPortex {
    /**
     * Call API
     */
    call<T>(path: string, options?: PortexRequestOptions): Promise<PortexResponse<T>>;
    /**
     * Initialize
     */
    init(): Promise<VerifyResult>;
    /**
     * Get start param
     */
    getStartParam(): string;
    /**
     * Check if user is verified
     */
    readonly isVerified: boolean;
    /**
     * Telegram Web App
     * @see https://core.telegram.org/bots/webapps
     */
    webApp: WebApp;
}